Refactor agent chat view cleanups
